+++ openstack/cbr/v3/policies/delete.go
@@ -1,0 +1,8 @@
+package policies
+
+import "github.com/opentelekomcloud/gophertelekomcloud"
+
+func Delete(client *golangsdk.ServiceClient, id string) (err error) {
+	_, err = client.Delete(client.ServiceURL("policies", id), nil)
+	return
+}
