+++ openstack/cbr/v3/policies/testing/requests_test.go
@@ -3,13 +3,14 @@ package testing
 import (
 	"testing"
 
+	golangsdk "github.com/opentelekomcloud/gophertelekomcloud"
 	"github.com/opentelekomcloud/gophertelekomcloud/openstack/cbr/v3/policies"
 	th "github.com/opentelekomcloud/gophertelekomcloud/testhelper"
 	fake "github.com/opentelekomcloud/gophertelekomcloud/testhelper/client"
 )
 
 func TestCreateV3PolicyMarshall(t *testing.T) {
-	res, err := createOpts.ToPolicyCreateMap()
+	res, err := golangsdk.BuildRequestBody(createOpts, "policy")
 	th.AssertNoErr(t, err)
 	th.AssertJSONEquals(t, expectedRequest, res)
 }
@@ -19,7 +20,7 @@ func TestCreateV3Policy(t *testing.T) {
 	defer th.TeardownHTTP()
 	handlePolicyCreation(t)
 
-	actual, err := policies.Create(fake.ServiceClient(), createOpts).Extract()
+	actual, err := policies.Create(fake.ServiceClient(), createOpts)
 	th.AssertNoErr(t, err)
 	th.AssertDeepEquals(t, expectedCreateResponseData, actual)
 }
@@ -29,7 +30,7 @@ func TestDeleteV3Policy(t *testing.T) {
 	defer th.TeardownHTTP()
 	handlePolicyDeletion(t)
 
-	err := policies.Delete(fake.ServiceClient(), "d32019d3-bc6e-4319-9c1d-6722fc136a22").ExtractErr()
+	err := policies.Delete(fake.ServiceClient(), "d32019d3-bc6e-4319-9c1d-6722fc136a22")
 	th.AssertNoErr(t, err)
 }
 
@@ -39,7 +40,7 @@ func TestUpdateV3Policy(t *testing.T) {
 	handlePolicyUpdate(t)
 
 	updateId := "cbb3ce6f-3332-4e7c-b98e-77290d8471ff"
-	actual, err := policies.Update(fake.ServiceClient(), updateId, updateOpts).Extract()
+	actual, err := policies.Update(fake.ServiceClient(), updateId, updateOpts)
 	th.AssertNoErr(t, err)
 	th.AssertDeepEquals(t, expectedCreateResponseData, actual)
 }
